refactor(sponsor-reports): move query builders to actions layer, fold in toOrderParam
